Update patch set 3
Patch Set 3: Looks good to me, but someone else must approve

No upstream fix for ContextImpl::databaseList yet.  But the fix for ContextImpl::deleteDatabase is working appropriately.

I pruned out the redundant parts in PS3 and tested this, appears to be behaving appropriately.  I think it's ready, if anyone else wants to give it a try.

Patch-set: 3