Update patch set 1
Patch Set 1:

Michael,

Thanks for the tip. I assume these are merged into the CM repositories. If so, this doesn't appear to completely solve the issue.

I've implemented the sdcard support correctly according to the Android storage examples for a primary non-emulated storage. First boot is absolutely fine and SD card works, but if for any reason the system_server crashes, a "hot reboot" results in the MediaProvider not being able to determine the UUID, causing the media service to never work until a full reboot.

I can also confirm this issue on the smultron using one of your recent CM11 releases. The issue can be easily reproduced via "killall system_server" and waiting for the system to reload. It will show the "android.process.media has stopped" errors and the reason is seen in the log (the UUID cannot be determined).

Patch-set: 1