Update patch set 3
Patch Set 3:

Hi Danny,
thanks for the explanation.
By running the following query I got a count of 987 records:

select count(*) from audio_meta where _data IS NULL;

I'll delete the media provider folder and let the system to recreate it again and check if I have a different result.

However, there are other users that are facing Eleven crashes when launching it, so it could surely useful to have a more robust method that prevents null pointer exceptions.

Patch-set: 3