Update patch set 1
Patch Set 1: Code-Review+1

Yes, it should always try if getFatVolumeId fails. The crux of the problem is that the Media Provider only supports 32-bit volume IDs. Proper UUIDs are 128-bits long so they will obviously be truncated but that doesn't matter. All that's needed is a value that's reasonably unique to that volume on that device. This is unlikely to be required elsewhere as a proper 128-bit value should be used wherever possible. We would probably use one here if backwards compatibility weren't an issue.

For the record, I have tried this patch and it works.

Patch-set: 1