Revert "MediaProvider: Fix where condition of delete queries"
Placeholders can point to a specific item of the arguments array.
When counting the number of arguments required to prepare a query,
items that point to the same argument are considered only once.
SQLiteProgram allows to provide an array that contains at most a
number of items equals to the number of arguments necessary for
the query. Therefore this change breaks delete queries in which
placeholders point to specific arguments.

This reverts commit fadf365acc9794897331e3896777df0751adc006.

Change-Id: Iefb683fb649008f0d90e23f11d0e06f90cda0727