Update patch set 2
Patch Set 2:

> This change causes some applications (ie; mxplayer) to require
 > sdcard authentication each time a file is delete.  Reverting this
 > commit fixes the issue.

I have been having issues deleting files and folders on an SD card beginning with a build compiled June 22, both with Solid Explorer and the in-built File Manager. While I cannot confirm it for sure, after examining the discussions here it seems quite possible that this commit is indeed the culprit. Just throwing this out there as additional confirmation of the issues probably doesn't hurt, and I'd be happy to see this fixed by next week's build.

Patch-set: 2