Update patch set 2
Patch Set 2: Verified+1 Code-Review+2

> I suggest we revert 5835a6fccf5fec1a2829e1f398416d8d10d8916e too
 > for now.

I don't disagree with this. That commit is probably causing more harm than good. I'm not sure when you hit the issue it tries to fix, but I don't really like that we can delete random data because the parent is missing in the database. We can always wait for a solution from Google now that they are aware of the problem, if we also find that the issue fixed by that commit is more an annoyance than a serious bug.

Patch-set: 2
Reviewer: Gabriele M <13897@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Verified=+1
Label: Code-Review=+2