Update patch set 2
Patch Set 2:

I perfectly understand that it's preferably to not break CTS, and if we were actually making a brand new device with lineage on it, then the situation would be completely different. But it is what it is. We're not, and if it stands between user experience and passing some CTS, I'll choose the former.

But as I said, if anyone has another solution, I'm all for it. I'm sure someone clever than me can figure out how to trigger a media scan after media is saved to external storage.

Patch-set: 2
Reviewer: Gerrit User 444 <444@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>