Update patch set 2
Patch Set 2:

> Without GAPPS preloaded: It does not work
> With GAPPS preloaded: It does work

Hmm, that's odd.

There's a patch sitting on AOSP gerrit, which I thought might have fixed it, but it made no difference here. Tested it as is on 17.1, and modified it for 18.1, with same result on both.

https://android-review.googlesource.com/c/platform/packages/providers/MediaProvider/+/1282810

Patch-set: 2
Reviewer: Gerrit User 444 <444@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>