Update patch set 2
Patch Set 2:

I think that if the issues caused by the earlier attempt to make Snap use mediastore was addressed, that would be a solution. I see that Google updated for latest API and mediastore in Camera2 in the master branch, in case that would help doing it for Snap. However, it seems like CAF hasn't been as good as Google to keep the storage stuff updated through the years. So it will likely be more work.

Patch-set: 2
Reviewer: Gerrit User 444 <444@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>