Update patch set 1
Patch Set 1:

It also builds, runs, and fixes the bug.

So two things:

a) What is the correct way of submitting this cherry-pick?

b) There are two other commits related to redaction ranges, which were made shortly afterwards:
https://android.googlesource.com/platform/packages/providers/MediaProvider/+/4b10f188ba319f9a6f45f11ab671611df66552a6
https://android.googlesource.com/platform/packages/providers/MediaProvider/+/e8882cec1d79545dbcb6303c0214d23878a96761
I'm not sure if would make sense to pick those as well (the associated bug reports are not accessible for me, so I have no idea what the implications are).

Patch-set: 1