Move markMediaAsFavorite's logic from MediaStore to MediaProvider
The content values retrieved from extras in the MediaProvider can have
values preset by the calling app and can affect other fields set.
Instead of retrieving the content values, we only retrieve a boolean
field that says whether the is_favorite is to be set to true or false.
All processing logic is shifted to MediaProvider.

Bug: 417194323
Test: EXEMPT Bugfix
Flag: EXEMPT Bugfix
Change-Id: I0227139b0a04efcb7456557fbf2d47996c017bad