Add missing permission check for albumart URI
The current permission check tests for album_art table, but doesn't deny
if there is no permission. If we start denying, we break the second
usage of these URIs which is redirection to thumbnail of an album, so
add additional check to allow this. These two different data systems do
not really conflict, as both permission to album_art and permission to
album can only be given by app which has broad audio permission.

Change-Id: I464ca89ba32a52f507563fd128d27cf50959ab34